webapp/requirements.txt
# Web Framework
flask>=3.0,<4
flask-socketio>=5.3,<6
flask-login>=0.6,<1
flask-sqlalchemy>=3.1,<4

# WebAuthn (passkey support)
py-webauthn>=2.0,<3

# WSGI Server
gunicorn>=21.0,<24

# Existing project dependencies
yfinance>=0.2.0
pandas>=1.5.0
numpy>=1.23.0
matplotlib>=3.6.0
Pillow>=9.0.0
openpyxl>=3.0.0
